refactor: add flag to force a refresh of the cache
