+++ src/routeProposal/index.ts
@@ -1,6 +1,12 @@
 import { filterPoolsOfInterest, filterHopPools } from './filtering';
 import { calculatePathLimits } from './pathLimits';
-import { SwapTypes, NewPath, PoolDictionary, SubgraphPoolBase } from '../types';
+import {
+    SwapOptions,
+    SwapTypes,
+    NewPath,
+    PoolDictionary,
+    SubgraphPoolBase,
+} from '../types';
 
 export class RouteProposer {
     processedDataCache: Record<
@@ -16,19 +22,17 @@ export class RouteProposer {
         tokenOut: string,
         swapType: SwapTypes,
         pools: SubgraphPoolBase[],
-        maxPools: number,
-        useProcessCache = true,
-        currentBlockTimestamp = 0
+        swapOptions: SwapOptions
     ): { pools: PoolDictionary; paths: NewPath[] } {
         if (pools.length === 0) return { pools: {}, paths: [] };
 
         // If token pair has been processed before that info can be reused to speed up execution
         const cache = this.processedDataCache[
-            `${tokenIn}${tokenOut}${swapType}${currentBlockTimestamp}`
+            `${tokenIn}${tokenOut}${swapType}${swapOptions.timestamp}`
         ];
 
-        // useProcessCache can be false to force fresh processing of paths/prices
-        if (useProcessCache && !!cache) {
+        // forceRefresh can be set to force fresh processing of paths/prices
+        if (!swapOptions.forceRefresh && !!cache) {
             // Using pre-processed data from cache
             return {
                 pools: cache.pools,
@@ -43,8 +47,8 @@ export class RouteProposer {
             poolsList,
             tokenIn,
             tokenOut,
-            maxPools,
-            currentBlockTimestamp
+            swapOptions.maxPools,
+            swapOptions.timestamp
         );
         const [filteredPoolsDict, pathData] = filterHopPools(
             tokenIn,
@@ -54,15 +58,12 @@ export class RouteProposer {
         );
         const [paths] = calculatePathLimits(pathData, swapType);
 
-        // Update cache if used
-        if (useProcessCache) {
-            this.processedDataCache[
-                `${tokenIn}${tokenOut}${swapType}${currentBlockTimestamp}`
-            ] = {
-                pools: filteredPoolsDict,
-                paths: paths,
-            };
-        }
+        this.processedDataCache[
+            `${tokenIn}${tokenOut}${swapType}${swapOptions.timestamp}`
+        ] = {
+            pools: filteredPoolsDict,
+            paths: paths,
+        };
 
         return { pools: filteredPoolsDict, paths };
     }
